include: remove irrelevant comments, and cast 868mhz to uint32_t

+++ include/lora.h
@@ -22,12 +22,11 @@
 // will ultimately depend on the board you have
 // some frequencies may be subject to regulations
 #define FREQ_433_MHZ ((uint32_t)433E6)
+// technically the 868 frequency is 868E6
+// however in some documentation ive seen 8681E5 used
+// and this seems to work better
 // #define FREQ_868_MHZ ((uint32_t)868E6)
-// should we do this instead?? not sure
-// there is technically a difference, however
-// from what i can tell there doesnt seem to make
-// a major difference at least in small trials
-#define FREQ_868_MHZ 8681E5
+#define FREQ_868_MHZ ((uint32_t)8681E5)
 #define FREQ_915_MHZ ((uint32_t)915E6)
 
 #define REG_FIFO 0x00
